Reuse system state labels
They are the same for chargers
and inverter/chargers
