+++ src/app/components/InverterCharger.js
@@ -8,6 +8,7 @@ import MqttWriteValue from "../mqtt/MqttWriteValue"
 import SelectorButton from "./SelectorButton"
 
 import { SYSTEM_MODE } from "../utils/constants"
+import SystemState from "./SystemState"
 
 const getTopics = (portalId, vebusInstanceId) => {
   return {
@@ -25,30 +26,15 @@ const systemModeFormatter = value => {
   return "--"
 }
 
-const systemStateFormatter = value => {
-  if (value == 0) return "Off"
-  if (value == 1) return "Low power"
-  if (value == 2) return "VE.Bus Fault condition"
-  if (value == 3) return "Bulk charging"
-  if (value == 4) return "Absorption charging"
-  if (value == 5) return "Float charging"
-  if (value == 6) return "Storage mode"
-  if (value == 7) return "Equalisation charging"
-  if (value == 8) return "Passthru"
-  if (value == 9) return "Inverting"
-  if (value == 10) return "Assisting"
-  if (value == 256) return "Discharging"
-  if (value == 257) return "Sustain"
-  return "--"
-}
-
 const InverterCharger = ({ activeMode, state, modeIsAdjustable, onModeSelected }) => {
   const systemMode = systemModeFormatter(activeMode)
   return (
     <div className="metric charger inverter-charger">
       <HeaderView icon={require("../../images/icons/multiplus.svg")} title="Inverter / Charger" child>
         <MetricValues>
-          <p className="text text--smaller">{systemStateFormatter(state)}</p>
+          <p className="text text--smaller">
+            <SystemState value={state} />
+          </p>
         </MetricValues>
       </HeaderView>
       <div className="charger__mode-selector">
